Origins/features/14 frontend create delete confirmation popup

+++ FrontEnd/ProjetoStarterFrontEnd/src/app/contract-holder/contract-holder.component.ts
@@ -8,6 +8,9 @@ import { GenericValidator } from '../Validations/GenericValidator';
 import { Observable } from 'rxjs';
 import { ActionButtonComponent } from '../action-button/action-button.component';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent, ConfirmDialogModel } from '../components/shared/confirmation-dialog/confirmation-dialog.component';
+import { MatDialog } from '@angular/material';
+
 
 @Component({
   selector: 'app-contract-holder',
@@ -18,6 +21,8 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
 
   rgMask = [/\d/, /\d/, '.', /\d/, /\d/, /\d/, '.', /\d/, /\d/, /\d/, '-', /[X0-9]/];
   cpfMask = [/\d/, /\d/, /\d/, '.', /\d/, /\d/, /\d/, '.', /\d/, /\d/, /\d/, '-', /\d/, /\d/];
+
+  public result: any;
 
 
 
@@ -36,60 +41,61 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
   showList: boolean = false;
   showAddresslist: boolean = false;
   showTelephonelist: boolean = false;
-    constructor(private chfb: FormBuilder, private http: HttpClient, private _snackBar: MatSnackBar, private location: Location) {
+  constructor(public dialog: MatDialog, private chfb: FormBuilder, private http: HttpClient, private _snackBar: MatSnackBar, private location: Location) {
 
   }
 
   message: number = 0;
   IndividualId: any = null;
+
+
   ngOnInit() {
     this.setup_gridData();
     this.setup_gridOptions();
     this.setup_form();
     this.setup_form();
-    
+
   }
 
   ngAfterViewInit() {
   }
 
   private handle_editUser(data: any) {
-    
+
     this.IndividualId = data.individualId;
     this.contractHolder.patchValue(data);
-    
-    
-    let telephoneControl =  this.contractHolder.controls.idTelephone as FormArray;
+
+
+    let telephoneControl = this.contractHolder.controls.idTelephone as FormArray;
     telephoneControl.controls.pop();
-    let a =0;
+    let a = 0;
     const hasMax = telephoneControl.length >= 5;
-      if (!hasMax) {
-        if (data.individualTelephones != ''){
-           for (a = 0; a < data.individualTelephones.length; a++){
-            
-            telephoneControl.push(this.chfb.group(data.individualTelephones[a]));
+    if (!hasMax) {
+      if (data.individualTelephones != '') {
+        for (a = 0; a < data.individualTelephones.length; a++) {
+
+          telephoneControl.push(this.chfb.group(data.individualTelephones[a]));
         }
-        }  
       }
-       
-      let addressControl = this.contractHolder.controls.idAddress as FormArray;
-      addressControl.controls.pop();
-      let b =0;
-      const hasMaxAddress = addressControl.length >= 3;
-      if (!hasMaxAddress) {
-        if (data.individualAddresses != '')
-        { 
-          for(b = 0 ;b < data.individualAddresses.length; b++) 
-               
-                 addressControl.push(this.chfb.group(data.individualAddresses[b]));
-                
-      }
-       
+    }
+
+    let addressControl = this.contractHolder.controls.idAddress as FormArray;
+    addressControl.controls.pop();
+    let b = 0;
+    const hasMaxAddress = addressControl.length >= 3;
+    if (!hasMaxAddress) {
+      if (data.individualAddresses != '') {
+        for (b = 0; b < data.individualAddresses.length; b++)
+
+          addressControl.push(this.chfb.group(data.individualAddresses[b]));
+
       }
 
     }
-    
-    private handle_deleteUser(data: any) {
+
+  }
+
+  private handle_deleteUser(data: any) {
     let json = JSON.stringify(this.contractHolder.value);
     let id = data.individualId;
     let httpOptions = {
@@ -97,18 +103,29 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
         'Content-Type': 'application/json'
       })
     };
+    const message = `Do you really want to delete this Contract Holder?`;
 
-    this.http.delete(`https://contractholderwebapi.azurewebsites.net/api/ContractHolder/${id}`). subscribe(data => this.setup_gridData(), error => this.openSnackBar(error.mensage),()=> this.openSnackBar('Titular deletado com sucesso') );      
-    
-    
+    const dialogData = new ConfirmDialogModel("Confirm Action", message);
 
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '375px',
+      panelClass: 'content-container',
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      this.result = dialogResult;
+      if (this.result == true) {
+        this.http.delete(`https://contractholderwebapi.azurewebsites.net/api/ContractHolder/${id}`).subscribe(data => this.setup_gridData(), error => this.openSnackBar(error.mensage), () => this.openSnackBar('Titular deletado com sucesso'));
+      }
+    });
   }
 
   unMaskValues(): void {
     let rg = this.contractHolder.controls.individualRG.value;
     rg = rg.replace(/\D+/g, '');
     this.contractHolder.controls.individualRG.setValue(rg);
-    
+
     let cpf = this.contractHolder.controls.individualCPF.value;
     cpf = cpf.replace(/\D+/g, '');
     this.contractHolder.controls.individualCPF.setValue(cpf);
@@ -132,7 +149,7 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
 
   changeMessageValue(): void {
     this.message = 1;
-  } 
+  }
 
   onSubmit(): void {
 
@@ -144,19 +161,19 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
         'Content-Type': 'application/json'
       })
     };
-      if (this.IndividualId == null) {
+    if (this.IndividualId == null) {
       this.http.post('https://contractholderwebapi.azurewebsites.net/api/contractholder', json, httpOptions).subscribe(response => this.load(), error => this.openSnackBar(error.message), () => this.openSnackBar("Titular cadastrado com sucesso"));
-  }
+    }
     else {
 
 
-    this.http.put(`https://contractholderwebapi.azurewebsites.net/api/ContractHolder/${this.IndividualId}`, json, httpOptions).subscribe(data => this.load(), error => this.openSnackBar(error.message), () => this.openSnackBar("Titular atualizado com sucesso"));
-}
-}
+      this.http.put(`https://contractholderwebapi.azurewebsites.net/api/ContractHolder/${this.IndividualId}`, json, httpOptions).subscribe(data => this.load(), error => this.openSnackBar(error.message), () => this.openSnackBar("Titular atualizado com sucesso"));
+    }
+  }
   openSnackBar(message: string): void {
     this._snackBar.open(message, '', {
       duration: 5000,
-      
+
     });
   }
 
@@ -175,7 +192,7 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
         addressType: ['', Validators.required],
         addressNumber: ['', [Validators.pattern(/^[0-9]+$/), Validators.maxLength(6)]],
         addressState: ['', [Validators.pattern(/^[[A-Z]+$/), Validators.maxLength(2), Validators.minLength(2)]],
-        addressNeighborhood: [ '', Validators.pattern(GenericValidator.regexSimpleName)],
+        addressNeighborhood: ['', Validators.pattern(GenericValidator.regexSimpleName)],
         addressCountry: ['', Validators.pattern(GenericValidator.regexSimpleName)],
         addressZipCode: ['', this.zipCodeValidation],
         addressCity: ['', Validators.pattern(GenericValidator.regexSimpleName)],
@@ -191,19 +208,19 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
 
     if (!hasMax) {
       telephoneControl.push(this.chfb.group({
-        telephoneNumber: ['', [GenericValidator.telephoneValidator(), ]],
+        telephoneNumber: ['', [GenericValidator.telephoneValidator(),]],
         telephoneType: ''
       }));
     }
     this.showTelephonelist = !this.showTelephonelist;
   }
- 
+
   handle_add_telphone($event: any) {
     let individualTelephonesControl = this.contractHolder.controls.individualTelephones as FormArray;
     $event.removeControl('telephoneId');
     individualTelephonesControl.push($event);
-  } 
-  
+  }
+
   handle_add_address($event: any) {
     let individualAddressesControl = this.contractHolder.controls.individualAddresses as FormArray;
     $event.removeControl('addressId');
@@ -281,7 +298,7 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
           onCellValueChanged: this.onCellEdit.bind(this),
           cellRenderer: (data) => {
             return data.value ? (new Date(data.value)).toLocaleDateString() : '';
-          }, 
+          },
         },
 
         {
@@ -334,14 +351,14 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
   private onCellEdit(params: any) {
   }
 
-  zipCodeValidation(control: AbstractControl): {[key: string]: boolean} | null {
+  zipCodeValidation(control: AbstractControl): { [key: string]: boolean } | null {
     let zipCodeNumber = control.value;
 
     zipCodeNumber = zipCodeNumber.replace(/\D+/g, '');
 
-    if(zipCodeNumber.length < 8)
-      return {"zipCodeIsTooShort": true};
-    
+    if (zipCodeNumber.length < 8)
+      return { "zipCodeIsTooShort": true };
+
     return null;
   }
 
@@ -350,4 +367,3 @@ export class ContractHolderComponent implements OnInit, AfterViewInit {
   }
 
 }
-
