+++ FrontEnd/ProjetoStarterFrontEnd/src/app/individual-list/individual-list.component.ts
@@ -5,6 +5,8 @@ import "ag-grid-enterprise";
 import { ActionButtonComponent } from '../action-button/action-button.component';
 import { ActionButtonBeneficiariesComponent } from '../action-button-beneficiaries/action-button-beneficiaries.component';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent, ConfirmDialogModel } from '../components/shared/confirmation-dialog/confirmation-dialog.component';
+import { MatDialog, MatDialogConfig } from '@angular/material';
 
 @Component({
   selector: 'app-individual-list',
@@ -13,7 +15,8 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 })
 export class IndividualListComponent implements OnInit {
 
-
+  public result: any;
+  individual: any;
   detailCellRendererParams;
   gridApi;
   gridColumApi;
@@ -22,29 +25,54 @@ export class IndividualListComponent implements OnInit {
   gridOptions: GridOptions;
   load_failure: boolean;
 
-  constructor(private http: HttpClient, private _snackBar: MatSnackBar) { }
+  constructor(public dialog: MatDialog, private http: HttpClient, private _snackBar: MatSnackBar) { }
 
   ngOnInit() {
     this.setup_gridData();
     this.setup_gridOptions();
     this.paginationPageSize = 50;
+
+    this.individual = this.http.get<Array<any>>('https://beneficiarieswebapi.azurewebsites.net/api/Beneficiary/Individuals');
   }
 
   private handle_editUser(data: any) {
     //this.contractform.patchValue(data);
-    }
-  
+  }
+
   private handle_deleteUser(data: any) {
     console.log(data);
     const id = data.beneficiaryId;
     console.log(id);
-    this.http.delete(`https://beneficiarieswebapi.azurewebsites.net/api/Beneficiary/${id}`).subscribe(response => this.setup_gridData(), error => this.openSnackBar(error.message), () => this.openSnackBar("Beneficiário removido com sucesso"));
+    const message = `Do you really want to delete this Beneficiary?`;
+
+    const dialogConfig = new MatDialogConfig();
+
+
+    const dialogData = new ConfirmDialogModel("Confirm Action", message);
+
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    dialogConfig.hasBackdrop = true;
+
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '375px',
+      panelClass: 'content-container',
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      this.result = dialogResult;
+      if (this.result == true) {
+        this.http.delete(`https://beneficiarieswebapi.azurewebsites.net/api/Beneficiary/${id}`).subscribe(response => this.setup_gridData(), error => this.openSnackBar(error.message), () => this.openSnackBar("Beneficiário removido com sucesso"));
+      }
+    });
   }
-    
+
   openSnackBar(message: string): void {
     this._snackBar.open(message, '', {
       duration: 5000,
-      
+
     });
   }
 
@@ -93,7 +121,7 @@ export class IndividualListComponent implements OnInit {
           filter: true,
           cellRenderer: (data) => {
             return data.value ? (new Date(data.value)).toLocaleDateString() : '';
-          }, 
+          },
           onCellValueChanged:
             this.onCellEdit.bind(this)
         },
@@ -131,6 +159,6 @@ export class IndividualListComponent implements OnInit {
 
   private onRowSelected(event: RowSelectedEvent) {
     const { data } = event;
-   
+
   }
 }
